refactor: Dvide API-related files according to purpose
